requirements.txt
asgiref==3.2.7
autopep8==1.5.2
boto3==1.13.14
botocore==1.16.14
Django==3.0.5
django-filter==2.2.0
django-phone-field==1.8.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
